review comments, and tests restructure
