+++ Dockerfile
@@ -47,12 +47,12 @@ ARG FEE_PAYER_TESTNET_MNEMONIC
 ARG FEE_PAYER_MAINNET_MNEMONIC
 
 # NPM environment variables
-ENV NPM_CONFIG_LOGLEVEL ${NPM_CONFIG_LOGLEVEL}
-ENV PORT ${PORT}
+ENV NPM_CONFIG_LOGLEVEL=${NPM_CONFIG_LOGLEVEL}
+ENV PORT=${PORT}
 
 # App-specific environment variables
-ENV FEE_PAYER_TESTNET_MNEMONIC ${FEE_PAYER_TESTNET_MNEMONIC}
-ENV FEE_PAYER_MAINNET_MNEMONIC ${FEE_PAYER_MAINNET_MNEMONIC}
+ENV FEE_PAYER_TESTNET_MNEMONIC=${FEE_PAYER_TESTNET_MNEMONIC}
+ENV FEE_PAYER_MAINNET_MNEMONIC=${FEE_PAYER_MAINNET_MNEMONIC}
 
 # Specify default port
 EXPOSE ${PORT}
