+++ tests/did/04-negative.spec.ts
@@ -1,35 +1,9 @@
-import { test, expect } from '@playwright/test';
+import { expect, test } from '@playwright/test';
 import { toString, fromString } from 'uint8arrays';
-import * as dotenv from 'dotenv';
-import { assert } from 'console';
+import { getDidDocument } from 'fixtures';
 
-dotenv.config();
-
-const pub_key_base_64 = process.env.TEST_PUBLIC_KEY;
-
-assert(pub_key_base_64, 'TEST_PUBLIC_KEY is not defined');
-
-const pubKeyHex = toString(fromString(pub_key_base_64 as string, 'base64pad'), 'base16');
-
-let didPayload;
 let indyDid = 'did:indy:sovrin:WRfXPg8dantKVubE3HX8pw';
 let deactiveDid = 'did:cheqd:testnet:ca9ff47c-0286-4614-a4be-8ffa83911e09';
-
-test('did-document. Generate the payload', async ({ request }) => {
-	const payload = await request.get(
-		`/1.0/did-document?verificationMethod=JsonWebKey2020&methodSpecificIdAlgo=uuid&network=testnet&publicKeyHex=${pubKeyHex}`
-	);
-
-	expect(payload.status()).toBe(200);
-
-	const body = await payload.json();
-	expect(body.didDoc).toBeDefined();
-	expect(body.key).toBeDefined();
-	expect(body.key.kid).toBeDefined();
-	expect(body.key.publicKeyHex).toBeDefined();
-
-	didPayload = body.didDoc;
-});
 
 test('did-create. wrong didDocument', async ({ request }) => {
 	const payload = await request.post('/1.0/create', {
@@ -52,6 +26,7 @@ test('did-create. wrong didDocument', async ({ request }) => {
 });
 
 test('did-update. invalid did', async ({ request }) => {
+	const didPayload = getDidDocument();
 	const payload = await request.post(`/1.0/update`, {
 		data: {
 			did: indyDid,
@@ -70,6 +45,7 @@ test('did-update. invalid did', async ({ request }) => {
 });
 
 test('did-update. Send deactivated did', async ({ request }) => {
+	const didPayload = getDidDocument();
 	const payload = await request.post(`/1.0/update`, {
 		data: {
 			did: deactiveDid,
@@ -88,6 +64,7 @@ test('did-update. Send deactivated did', async ({ request }) => {
 });
 
 test('did-update. Send wrong operation', async ({ request }) => {
+	const didPayload = getDidDocument();
 	const payload = await request.post(`/1.0/update`, {
 		data: {
 			didDocument: [didPayload],
